Day 25-26 Work Session 1
In today's work session, I looked over the general guidelines for Lab 4, as well as both of the websites. I ended up choosing the Project Euler Problems, and I successfully got through 2 of them, but I was stuck problem 3. The concepts for problem 1 and 2 were quite easy, and after a bit of revision I was able to completely write the code. The third question, however, was a bit harder to do. I initially though of finding all the factors first, then finding ones that are prime. However, this was an extremely inefficient way of solving this problem. I though that my code was stuck in an infinite loop at first, but I realized that the number was simply too large to calculate. After looking online, some people have found solutions, though I didn't really understand what they are doing. I hope to discuss this with my peers in class.