Update CSLab4.py
In this work session, I continued to work on the problems on Project Euler. The problems were increasing in difficulty, and I started today off by finishing problem 3. Last time I tried to brute force it, but obviously it didn't work. I realized that an efficient way would be to find each factor, but after each one is identified, divide that by the original number so the initial "n" value would be smaller, and decreases with each factor found. ultimately, I got the right answer in less than a few seconds, meaning that the program no longer needed to test every single possible factor. Problem 4 was equally as challenging to me. To find the largest palindrome number, I used the range functions in for loops to identify the x and y number (generally). obviously, the product of x and y would form the number we want. If that number then passes the test of palindrome through this line: "if StringProduct == StringProduct[::-1]:", then it is the answer. I had the most trouble working on question 5. I am asked to find the LCM of numbers between 1-20. I initially started to write a brute force algorithm, but I stopped after realizing it wouldn't work efficiently. I spent an additional 30 mins looking for alternate methods, but I couldn't understand them. I hope that I can talk to my peers in the future, and see how they tackled this problem.