Updated CS Lab4.py (Jan 10)
In this work session, I achieved more than I initially anticipated. I continued to work on question 5 for the first 30 minutes, but I was unable to think of an effective solution except for the brute for algorithm, so I decided to move on and ask my peers tomorrow in class. I completely finished problems 6, 8, and 9 today. Problem 6 was quite straight forward. I essentially used for loops to calculate the “sum of squares” as well as the “square of sums”, and in the end subtracted them to get my final answer. Number 8 was a bit challenging to me, and took me some time to complete. While its minor, I found it hard to copy and paste the fill sequence of numbers and format it correctly onto the code. Besides that, however, I spent a lot of time figuring out the most efficient way to calculate chunks of 14 digits throughout the entire sequence of numbers. In the end, I implemented a for loop that runs through the entire sequence, and another for loop inside that multiplies each of the 14 numbers in the set together. The rest of the problem was easy since it only requires substituting variables. For number 9, I couldn’t think of a super efficient way to figure out the answer to I believe that my current solution is a brute force algorithm. I used 2 for loops to calculate the a and b values, and used simple maths to solve for c. inside the same for loop, I checked if the multiple of a, b, and c are equal to exactly 1000. If not, the loop continues the find new values. If it passes, then it will print the answer. I also encountered some issues with number 7. Specifically, I didn’t know how to find all of the prime numbers. I wish to discuss that in class with my peers as well as number 5.