Updated CSLab4.py Jan 17
I was quite satisfied with my work in today’s work session. While it took be at least 30 minutes of planning and thinking, I finally figured out how to do problem 5 of projectEuler. One thing that I realized was that we could actually shrink the size of the list. For instance, since the factors of 15 are also in the list, “1,3,5”, we can eliminate those so that the code can run less times overall. However, that wasn’t the greatest obstacle for this problem. I created a for loop to run through every possible “LCM” of the numbers in the set. The for loop runs up to 999999999 (a reasonably large number), starts at 20 (the LCM must be at least equal to the largest number in the list), and has a step size of 20 (the LCM must be a multiple of 20). I discovered “all()”, which will return either true or false. If the conditions inside passes, the code will continue to run. That said, with all the numbers passed through, the resulting value would be all of the numbers common factors under “999999999”. At first, it kept returning larger multiples. It took me a while, but I realized I could just stop it by using “exit()”. After finishing the problem that I was stuck on for 2 work session, I moved on to problem 7. Just like question 5, I impliented “all()”, and was able to find all the prime numbers under 10001. However, I have yet to write out the section that would return the largest value in the list. Perhaps I could use one of the sort algorithms we tried out in class and obtain the largest prime number in the list.